Ensure modules with same name and different paths have unique keys
