+++ ward/collect.py
@@ -10,6 +10,7 @@ from typing import Any, Callable, Generator, Iterable, List, Set
 
 from ward.models import WardMeta
 from ward.testing import Test, anonymous_tests
+from ward.util import get_absolute_path
 
 
 def is_test_module(module: pkgutil.ModuleInfo) -> bool:
@@ -55,8 +56,8 @@ def load_modules(modules: Iterable[pkgutil.ModuleInfo]) -> Generator[Any, None, 
 def get_tests_in_modules(modules: Iterable) -> Generator[Test, None, None]:
     for mod in modules:
         mod_name = mod.__name__
-        # Collect anonymous tests from the module
-        anon_tests: List[Callable] = anonymous_tests[mod_name]
+        mod_path = get_absolute_path(mod)
+        anon_tests: List[Callable] = anonymous_tests[mod_path]
         if anon_tests:
             for test_fn in anon_tests:
                 meta: WardMeta = getattr(test_fn, "ward_meta")
