+++ ward/run.py
@@ -17,8 +17,8 @@ from ward.collect import (
 from ward.config import read_config_toml
 from ward.rewrite import rewrite_assertions_in_tests
 from ward.suite import Suite
-from ward.terminal import SimpleTestResultWrite
-from ward.util import get_exit_code, find_project_root
+from ward.terminal import SimpleTestResultWrite, get_exit_code
+from ward.util import find_project_root
 
 init()
